Removed several keys and keyrefs that used multiple fields in a key, as they were not behaving as expected.
Where a selector in a key definition may select more than one element, the key for a matched element may not match more than one node. So, a key for the combination of valid attribute names for a search program can not be constructed, since the selector can select more than one search program, but the key for the name of the attribute will fail, since there there are multiple attributes per program. Doing this correctly would require storing (redundantly) the program name as an atribute of a score name, or restructuring the XML to store actual PSM and peptide score under the score type elements instead of under peptides.